Fix #10874: Ensure View Bar Order Matches Navigation Bar
Fixed an issue where the order of the view bar did not match the
navigation bar, even when deleting existing views, creating
new views, or reordering them.

Created a function on ViewPickerListContent to update positions
and modified the handleDragEnd to update the positions of the
views.
Modified on useCreateViewFromCurrentView to update the
positions when a new view is created.

With both this modifications the view order remains consistent
across both components.